feat: add async action
